feat: add CVE year-tracking metrics and Prometheus metric enhancements (#297)
* Add CVE year-tracking and Prometheus metric enhancements and fixes the total cve counter

Introduced `CveCounterPerYear` to manage CVE counts per year, updating Prometheus metrics for detailed tracking. A new file: `cve-count-per-year.json` will be created, in which the number of CVEs per year is recorded. Added gauges for sync time, along with optimized cache handling and error logging. Fixes the total CVE counts due to the restructuring of the cache regarding memory behavior (#288).
 Updated docs to correct metric writer configuration details. Typo: (METRICS_WRITE_FORMAT -> METRICS_WRITER_FORMAT)

* Use logging instead of throwing CacheException. (#297)

Replaced the thrown `CacheException` with a warning log message when unable to read `count.json`. This ensures the application does not fail but logs the issue for further investigation and at the end, the file is recreated.

* Add link for cve-counter-per-year.json and the version on the homepage (#297)

The version number of the mirror is now displayed on the homepage. Additionally, the file `cve-counter-per-year.json` has been added as a hyperlink for download. The JSON file is now written with pretty printing for better readability. The correct `README.md` has been updated to fix the typo: (METRICS_WRITE_FORMAT -> METRICS_WRITER_FORMAT).
During validation, the `cve-count-per-year.json` is now taken into account.

* Rename log message to reference accurate JSON file name (#297)

The log message previously referred to "count.json" but the actual file is "cve-count-per-year.json". This update corrects the log message to avoid confusion and improve clarity.